+++ config/config.exs
@@ -7,6 +7,5 @@ end
 if Mix.env() == :test do
   config :ua_inspector,
     database_path: Path.expand("../test/fixtures", __DIR__),
-    ets_cleanup_delay: 10,
-    skip_download_readme: true
+    ets_cleanup_delay: 10
 end